% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{calculateMeanExposure}
\alias{calculateMeanExposure}
\title{Calculate mean exposure.}
\usage{
calculateMeanExposure(selNodes, adjMat, statusMat, column = 1,
  bLog = FALSE)
}
\arguments{
\item{selNodes}{Indices of the specified nodes.}

\item{adjMat}{Adjacency matrix for the network.}

\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{column}{Column to analyze.}

\item{bLog}{Whether to print infective links.}
}
\value{
A vector with two elements: number of incoming links, number of incoming links from infectious units (status > 0).
}
\description{
Calculates mean exposure from incoming links for a set of node.
}
\details{
Implemented via matrix multiplications.
}
\author{
Mikhail Churakov
}
