% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{calculateNodeExposure}
\alias{calculateNodeExposure}
\title{Calculate node exposure.}
\usage{
calculateNodeExposure(node, adjMat, statusMat, column = 1)
}
\arguments{
\item{node}{Index of the specified node.}

\item{adjMat}{Adjacency matrix for the network.}

\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{column}{Column to analyze.}
}
\value{
Statuses of incoming links for the specified node.
}
\description{
Calculates incoming links for a node.
}
\author{
Mikhail Churakov
}
