% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{confMat}
\alias{confMat}
\title{Print probabilities of changing infection status.}
\usage{
confMat(statusMat, cols = c(1, 2))
}
\arguments{
\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{cols}{Columns (i.e. timepoints) to analyse. Only the first two will be used.}
}
\value{
Returns the matrix with probabilities.
}
\description{
Probabilities of switching status between two timepoints are calculated and printed.
}
\author{
Mikhail Churakov
}
