% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{createPPPList}
\alias{createPPPList}
\title{Create a list of PPP objects}
\usage{
createPPPList(statusMat, coords, cols = c(1, 2), na.rm = FALSE)
}
\arguments{
\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{coords}{Coordinates of units.}

\item{cols}{Columns (i.e. timepoints) to analyse, by default the first two are used.}

\item{na.rm}{Whether remove NA values for each timepoint separately.}
}
\value{
List of PPP objects.
}
\description{
Status matrix and corresponding coordinates are used to create a list of PPP objects for further spatial analysis.
}
\author{
Mikhail Churakov
}
