% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{createRandomNetwork}
\alias{createRandomNetwork}
\title{Create a random network}
\usage{
createRandomNetwork(nv, ne, weights = rep(1, ne), bDirected = FALSE)
}
\arguments{
\item{nv}{Number of nodes (vertices).}

\item{ne}{Number of edges (links).}

\item{weights}{Weight of links (vector of length `ne`)}

\item{bDirected}{Whether to make the matrix symmetrical (i.e. biderectional links).}
}
\value{
Adjacency matrix. Non-zero element adjMat[i, j] indicates the weight of the link between nodes i and j.
}
\description{
Creates a random network with the specified number of nodes and links.
}
\author{
Mikhail Churakov
}
