% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{densityPPPlist}
\alias{densityPPPlist}
\title{Plot density of PPP objects}
\usage{
densityPPPlist(lst)
}
\arguments{
\item{lst}{List of PPP objects.}
}
\description{
Plots density of all PPP objects from the provided list.
}
\author{
Mikhail Churakov
}
