% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{logrrPPPlist}
\alias{logrrPPPlist}
\title{Plot log-ratio for PPP objects}
\usage{
logrrPPPlist(lst)
}
\arguments{
\item{lst}{List of PPP objects.}
}
\description{
Plots log-ratio [...] for each PPP object from the provided list.
}
\author{
Mikhail Churakov
}
