% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{meanExpFun}
\alias{meanExpFun}
\title{Calculate mean exposure.}
\usage{
meanExpFun(adjMat, statusMat, cols = 1:2, nsim = 1000, bLog = FALSE)
}
\arguments{
\item{adjMat}{Adjacency matrix for the network.}

\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{cols}{Column (i.e. timepoints) to analyse. Only one column will be used.}

\item{nsim}{Number of simulations.}

\item{bLog}{Whether to print extra info.}
}
\value{
List of observed and simulated mean exposures for NI and PS units: obsNI, obsPS, mexpNI, mexpPS.
}
\description{
Calculates mean exposure for NI and PS units.
}
\author{
Mikhail Churakov
}
