% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{permuteLinks}
\alias{permuteLinks}
\title{Permute links.}
\usage{
permuteLinks(adjMat)
}
\arguments{
\item{adjMat}{Adjacency matrix for the network.}
}
\value{
Adjacency matrix with permuted links.
}
\description{
Permutes links in the network.
}
\details{
Permutation is equivalent to creating a new adjustment matrix with the same number of nodes and edges.
}
\author{
Mikhail Churakov
}
