% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{plotDistr}
\alias{plotDistr}
\title{Plot distributions}
\usage{
plotDistr(statusMat, distMat, bTiff = FALSE, cols = c(1, 2))
}
\arguments{
\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{distMat}{Distance matrix.}

\item{bTiff}{Whether to generate TIFF file.}

\item{cols}{Column (i.e. timepoints) to analyse. Only one column will be used.}
}
\description{
.
}
\author{
Mikhail Churakov
}
