% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{plotHists}
\alias{plotHists}
\title{Plot histograms}
\usage{
plotHists(statusMat, distMat, cols = c(1, 2))
}
\arguments{
\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{distMat}{Distance matrix.}

\item{cols}{Column (i.e. timepoints) to analyse. Only one column will be used.}
}
\description{
.
}
\author{
Mikhail Churakov
}
