% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{plotMeanExp}
\alias{plotMeanExp}
\title{Plot mean exposure.}
\usage{
plotMeanExp(obsNI, obsPS, mexpNI, mexpPS, bEllipse = FALSE,
  elLevel = 0.95, quantileVals = c(0.05, 0.95), bShowPerc = TRUE,
  ...)
}
\arguments{
\item{obsNI}{Observed mean exposure for NI units.}

\item{obsPS}{Observed mean exposure for PS units.}

\item{mexpNI}{Simulated mean exposures for NI units.}

\item{mexpPS}{Simulated mean exposures for PS units.}

\item{bEllipse}{Whether to show ellipse containing a fraction of points.}

\item{elLevel}{Fraction of points that should be inside the ellipse (if bEllipse is TRUE).}

\item{quantileVals}{Vector of quantiles used to define significance levels.}

\item{bShowPerc}{Whether to show percentages of simulated values below the observed exposure.}

\item{...}{Additional graphic parameters.}
}
\description{
Plot mean exposure for NI and PS units.
}
\author{
Mikhail Churakov
}
