% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot network data.}
\usage{
plotNetwork(coords, adjMat, color = "black", ...)
}
\arguments{
\item{coords}{Coordinates of units.}

\item{adjMat}{Adjacency matrix for the network.}

\item{color}{Color of nodes (can be a vector).}

\item{...}{Additional graphic parameters.}
}
\description{
Makes a plot of a provided network.
}
\author{
Mikhail Churakov
}
