% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plotPPPlist}
\alias{plotPPPlist}
\title{Plot PPP objects}
\usage{
plotPPPlist(lst)
}
\arguments{
\item{lst}{List of PPP objects.}
}
\description{
Plot all PPP objects from the provided list.
}
\author{
Mikhail Churakov
}
