% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotStrainDistribution}
\alias{plotStrainDistribution}
\title{Plot strain distribution}
\usage{
plotStrainDistribution(statusMat, colors = 1:ncol(statusMat),
  exclude = c(), ...)
}
\arguments{
\item{statusMat}{A matrix with statuses (NA, 0, any positive integer) of each unit. Rownames indicate location names, colnames indicate names of timepoints (e.g. years).}

\item{colors}{Colors to be used.}

\item{exclude}{Strains that should be excluded.}

\item{...}{Additional graphic parameters.}
}
\description{
Plots a barplot with strain distributions for considered timepoints.
}
\author{
Mikhail Churakov
}
