% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{satscanPPPlist}
\alias{satscanPPPlist}
\title{Perform SaTScan test on PPP objects}
\usage{
satscanPPPlist(lst)
}
\arguments{
\item{lst}{List of PPP objects.}
}
\description{
Runs SaTScan test (implemented in smacpod package) on each PPP object from the provided list.
}
\author{
Mikhail Churakov
}
