% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatclustKfun}
\alias{spatclustKfun}
\title{Perform K-function test}
\usage{
spatclustKfun(lst, nsim = 10)
}
\arguments{
\item{lst}{List of PPP objects.}

\item{nsim}{Number of simulations. Default is 10.}
}
\description{
Plots results of K-function spatial clustering test.
}
\details{
K-function is measured for the observed and simulated spatial dictributions.
}
\author{
Mikhail Churakov (\email{mikhail.churakov@gmail.com}).
}
