% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatclustkNN}
\alias{spatclustkNN}
\title{Perform Cuzick-Edwards' kNN test}
\usage{
spatclustkNN(lst, nn = 20, nsim = 500, signifLevel = 0.05)
}
\arguments{
\item{lst}{List of PPP objects.}

\item{nn}{Number of nearest neighbours. Default is 20.}

\item{nsim}{Number of simulations. Default is 500.}

\item{signifLevel}{Significance level (p-value cut-off). Default is 0.05.}
}
\description{
Plots results of Cuzick-Edwards' kNN clustering test.
}
\author{
Mikhail Churakov
}
