% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{textPlot}
\alias{textPlot}
\title{Make a simple plot with a centered text}
\usage{
textPlot(text = "", ...)
}
\arguments{
\item{text}{Text for the plot.}

\item{...}{Additional graphic parameters.}
}
\description{
Makes an empty plot with a provided text in the middle.
}
\author{
Mikhail Churakov
}
